Change doubles to ints

+++ graphs/lib/models/sprite.dart
@@ -1,8 +1,8 @@
 class Sprite {
   static int _id = 0;
   late int id;
-  late double x;
-  late double y;
+  late int x;
+  late int y;
 
   Sprite({required this.x, required this.y}) {
     id = _id++;
